tools/data_enrichment/requirements.txt
python-dotenv>=1.0.1
requests>=2.32.3
beautifulsoup4>=4.12.3
duckduckgo-search>=6.2.11
psycopg[binary]>=3.2.1
